Full-cohort run: 1003 users, 4 LLMs, STM/MTM trained and reported
Scaling the pipeline from a 5-user smoke test to the full cohort exposed three
blockers, all fixed here:

- Captioning ran one image per forward pass (97 h for 165k images). scripts/
  caption_shard.py batches and shards across GPUs: 2.2 h at 21 img/s.
- Extraction asked for output_hidden_states, materializing every layer over the
  whole prompt (tens of GB per user at 128k-char corpora) and prefilling twice
  per chunk. It now taps layers with forward hooks during a single generate pass;
  scripts/validate_taps.py shows the result is bit-identical. This also fixes
  Gemma's layer count, which config-based detection read as 12 instead of 30.
- The 504-point grid ran serially on CPU. scripts/train_parallel.py routes wide
  nets to GPUs and narrow nets to single CPU threads across 48 workers: 5040 fits
  in 40 min.

device_map: auto is avoided throughout: sharding a model across this node's four
Blackwell cards yields numerically corrupt activations, so each LLM is pinned to
one GPU and Llama-3.3-70B runs 8-bit.

Results (5-fold CV, y_high = suicide >= 3, 132/1003 positives):
STM AUC 0.643 [0.617, 0.670], MTM AUC 0.674 [0.623, 0.725]. The models' own
zero-shot RISK verdicts sit near chance (0.52-0.58), so the signal is in the
hidden states rather than the verbalized answer.

Per-user hidden states (~900 MB) stay local; the posts, cohort, captions and
corpora needed to re-run the LLM stage are committed.

+++ src/ssr/represent/store.py
@@ -18,7 +18,8 @@ def save_user_reps(path: Path, arrays: dict[str, np.ndarray], meta: dict) -> Non
     path.parent.mkdir(parents=True, exist_ok=True)
     # np.savez_compressed appends .npz if missing — use a temp stem carefully
     tmp = path.parent / (path.stem + ".partial.npz")
-    payload = {k: np.asarray(v, dtype=np.float16) for k, v in arrays.items()}
+    # Keep float32: mid-layer bf16 activations often overflow float16 (→ inf/nan).
+    payload = {k: np.asarray(v, dtype=np.float32) for k, v in arrays.items()}
     meta_json = np.array([str(meta)])
     np.savez_compressed(tmp, __meta__=meta_json, **payload)
     # np may write exactly `tmp` already ending in .npz
